+++ src/app/admin/components/manage-products/manage-products.component.ts
@@ -1,8 +1,11 @@
 import {Component, OnInit} from '@angular/core';
-import {Observable} from 'rxjs';
 
+import {Observable} from 'rxjs';
+import {select, Store} from '@ngrx/store';
+
+import {AppState, getProductsCount, getProductsData, getProductsError} from '../../../core/+store';
+import * as ProductsActions from './../../../core/+store/products/products.actions';
 import {ProductModel} from '../../../products/models/product.model';
-import {ProductsService} from '../../../products/services/products.service';
 
 @Component({
   selector: 'app-manage-users',
@@ -10,13 +13,19 @@ import {ProductsService} from '../../../products/services/products.service';
   styleUrls: ['./manage-products.component.css']
 })
 export class ManageProductsComponent implements OnInit {
-  products$: Observable<ProductModel[]>;
+  products$: Observable<ReadonlyArray<ProductModel>>;
+  productsError$: Observable<Error | string>;
+  productsCount$: Observable<number>;
 
-  constructor(private productsService: ProductsService) {
+  constructor(private store: Store<AppState>) {
   }
 
   ngOnInit() {
-    this.products$ = this.productsService.getProducts();
+    this.products$ = this.store.pipe(select(getProductsData));
+    this.productsError$ = this.store.pipe(select(getProductsError));
+    this.productsCount$ = this.store.pipe(select(getProductsCount));
+
+    this.store.dispatch(new ProductsActions.GetProducts());
   }
 
 }