+++ src/app/orders/models/order.model.ts
@@ -2,6 +2,7 @@ import {CartItemModel} from '../../cart/models/cart-item.model';
 
 export class OrderModel {
   constructor(
+    public id: number = null,
     public products: CartItemModel[],
     public sum: number,
     public quantity: number,