+++ src/app/products/products.module.ts
@@ -2,9 +2,13 @@ import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 
+import {StoreModule} from '@ngrx/store';
+import {EffectsModule} from '@ngrx/effects';
+
 import {ProductComponent, ProductDetailsComponent, ProductListComponent, ProductReviewsComponent} from './components';
 import {SharedModule} from '../shared/shared.module';
 import {ProductsRoutingModule} from './products-routing.module';
+import {ProductsEffects, productsReducer} from '../core/+store/products';
 
 @NgModule({
   declarations: [
@@ -18,6 +22,8 @@ import {ProductsRoutingModule} from './products-routing.module';
     FormsModule,
     ProductsRoutingModule,
     SharedModule,
+    StoreModule.forFeature('products', productsReducer),
+    EffectsModule.forFeature([ProductsEffects]),
   ],
   exports: [
     ProductListComponent,